+++ src/test/java/com/artipie/helm/HelmSliceIT.java
@@ -21,79 +21,64 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+
 package com.artipie.helm;
 
+import com.artipie.asto.Key;
 import com.artipie.asto.Storage;
+import com.artipie.asto.blocking.BlockingStorage;
 import com.artipie.asto.memory.InMemoryStorage;
 import com.artipie.http.rs.RsStatus;
+import com.artipie.http.rs.StandardRs;
 import com.artipie.vertx.VertxSliceServer;
+import com.google.common.collect.Maps;
 import io.vertx.reactivex.core.Vertx;
 import io.vertx.reactivex.core.buffer.Buffer;
 import io.vertx.reactivex.ext.web.client.WebClient;
+import java.io.File;
 import java.io.IOException;
-import java.net.ServerSocket;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+import org.awaitility.Awaitility;
 import org.hamcrest.MatcherAssert;
+import org.hamcrest.collection.IsMapContaining;
+import org.hamcrest.collection.IsMapWithSize;
+import org.hamcrest.core.IsAnything;
 import org.hamcrest.core.IsEqual;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
+import org.hobsoft.hamcrest.compose.ComposeMatchers;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
+import org.junit.jupiter.api.io.TempDir;
 import org.slf4j.LoggerFactory;
-import org.testcontainers.Testcontainers;
-import org.testcontainers.containers.Container;
 import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.shaded.org.yaml.snakeyaml.Yaml;
 
 /**
- * Ensure that helm command line tool is compatible with this adapter.
+ * Push helm chart and ensure if index.yaml is generated properly.
  *
+ * @since 0.2
  * @checkstyle MethodBodyCommentsCheck (500 lines)
  * @checkstyle ClassDataAbstractionCouplingCheck (500 lines)
  * @checkstyle MagicNumberCheck (500 lines)
  * @checkstyle ExecutableStatementCountCheck (500 lines)
- * @since 0.2
  */
-@DisabledIfSystemProperty(named = "os.name", matches = "Windows.*")
 @SuppressWarnings("PMD.AvoidDuplicateLiterals")
-public final class HelmSliceIT {
+@DisabledIfSystemProperty(named = "os.name", matches = "Windows.*")
+public class HelmSliceIT {
 
-    /**
-     * The vertx.
-     */
-    private Vertx vertx;
-
-    /**
-     * The helm.
-     */
-    private HelmSliceIT.HelmContainer helm;
-
-    /**
-     * The web.
-     */
-    private WebClient web;
-
-    /**
-     * The turl.
-     */
-    private String turl;
-
-    /**
-     * The server.
-     */
-    private VertxSliceServer server;
-
-    @BeforeEach
-    public void before() throws URISyntaxException, IOException {
-        this.vertx = Vertx.vertx();
+    @Test
+    public void indexYamlIsCorrect(@TempDir final Path tmp) throws IOException, URISyntaxException {
+        final Vertx vertx = Vertx.vertx();
         final Storage fls = new InMemoryStorage();
         final int port = rndPort();
-        this.turl = String.format("http://host.testcontainers.internal:%d/", port);
-        Testcontainers.exposeHostPorts(port);
-        this.server = new VertxSliceServer(
-            this.vertx,
-            new HelmSlice(fls, this.turl),
+        final VertxSliceServer server = new VertxSliceServer(
+            vertx,
+            new HelmSlice(fls, String.format("http://localhost:%d/", port)),
             port
         );
         final byte[] tomcat = Files.readAllBytes(
@@ -104,77 +89,82 @@ public final class HelmSliceIT {
                     .toURI()
             )
         );
-        this.web = WebClient.create(this.vertx);
-        this.helm = new HelmSliceIT.HelmContainer()
-            .withCreateContainerCmdModifier(
-                cmd -> cmd.withEntrypoint("/bin/sh").withCmd("-c", "while sleep 3600; do :; done")
+        final WebClient web = WebClient.create(vertx);
+        try {
+            server.start();
+            final int code = web.post(port, "localhost", "/")
+                .rxSendBuffer(Buffer.buffer(tomcat))
+                .blockingGet()
+                .statusCode();
+            MatcherAssert.assertThat(
+                code,
+                new IsEqual<>(Integer.parseInt(RsStatus.OK.code()))
             );
-        this.server.start();
-        final int code = this.web.post(port, "localhost", "/")
-            .rxSendBuffer(Buffer.buffer(tomcat))
-            .blockingGet()
-            .statusCode();
-        if (code != Integer.parseInt(RsStatus.OK.code())) {
-            throw new IllegalStateException(
-                String.format("Received code non-200 code:%d", code)
+            final String adapter = new String(
+                new BlockingStorage(fls).value(new Key.From("index.yaml"))
             );
+            LoggerFactory.getLogger(HelmSliceIT.class).info(
+                "Generated by adapter index.yaml:\n{}",
+                adapter
+            );
+            Files.write(tmp.resolve("tomcat-0.4.1.tgz"), tomcat);
+            final HelmContainer helm = new HelmContainer()
+                .withCommand("repo", "index", ".")
+                .withWorkingDirectory("/home/")
+                .withFileSystemBind(tmp.toString(), "/home");
+            helm.start();
+            final File index = tmp.resolve("index.yaml").toFile();
+            Awaitility.await().atMost(5, TimeUnit.SECONDS).until(index::exists);
+            final String expected = new String(Files.readAllBytes(index.toPath()));
+            LoggerFactory.getLogger(HelmSliceIT.class).info(
+                "Generated by helm index.yaml:\n{}",
+                expected
+            );
+            MatcherAssert.assertThat(
+                Maps.difference(
+                    tomcatZeroEntry(expected),
+                    tomcatZeroEntry(adapter)
+                ).entriesDiffering(),
+                ComposeMatchers.compose(new IsMapWithSize<String, Object>(new IsEqual<>(2)))
+                    .and(new IsMapContaining<>(new IsEqual<>("urls"), new IsAnything<>()))
+                    .and(new IsMapContaining<>(new IsEqual<>("created"), new IsAnything<>()))
+            );
+        } finally {
+            web.close();
+            server.close();
+            vertx.close();
         }
-        this.helm.start();
-    }
-
-    @AfterEach
-    public void after() {
-        this.helm.stop();
-        this.web.close();
-        this.server.close();
-        this.vertx.close();
     }
 
     /**
-     * Helm add repo works.
-     *
-     * @throws IOException If fails
-     * @throws InterruptedException If fails
+     * Parse and return index.yaml.
+     * @param yaml The index.yaml file passed as string
+     * @return The first element in entries->tomcat
      */
-    @Test
-    public void helmRepoAddAndUpateWorks() throws IOException, InterruptedException {
-        exec(this.helm, "helm", "init", "--client-only", "--debug");
-        MatcherAssert.assertThat(
-            "helm repo add failed",
-            exec(this.helm, "helm", "repo", "add", "test", this.turl),
-            new IsEqual<>(0)
-        );
-        MatcherAssert.assertThat(
-            "helm repo update failed",
-            exec(this.helm, "helm", "repo", "update"),
-            new IsEqual<>(0)
-        );
-    }
-
-    private int exec(
-        final HelmSliceIT.HelmContainer helmc,
-        final String... cmd) throws IOException, InterruptedException {
-        final String joined = String.join(" ", cmd);
-        LoggerFactory.getLogger(EnsureIndexIsGoodITCase.class).info("Executing:\n{}", joined);
-        final Container.ExecResult exec = helmc.execInContainer(cmd);
-        LoggerFactory.getLogger(EnsureIndexIsGoodITCase.class)
-            .info("STDOUT:\n{}\nSTDERR:\n{}", exec.getStdout(), exec.getStderr());
-        final int code = exec.getExitCode();
-        if (code != 0) {
-            LoggerFactory.getLogger(EnsureIndexIsGoodITCase.class)
-                .error("'{}' failed with {} code", joined, code);
-        }
-        return code;
+    @SuppressWarnings("unchecked")
+    private static Map<String, Object> tomcatZeroEntry(final String yaml) {
+        return (
+            (ArrayList<Map<String, Object>>)
+                ((Map<String, Object>) new Yaml().<Map<String, Object>>load(yaml)
+                    .get("entries")).get("tomcat")
+        ).get(0);
     }
 
     /**
      * Obtain a random port.
-     *
      * @return The random port.
      */
-    private static int rndPort() throws IOException {
-        try (ServerSocket socket = new ServerSocket(0)) {
-            return socket.getLocalPort();
+    private static int rndPort() {
+        final Vertx vertx = Vertx.vertx();
+        final VertxSliceServer server = new VertxSliceServer(
+            vertx,
+            (line, headers, body) -> StandardRs.EMPTY
+        );
+        try {
+            return server.start();
+        } finally {
+            server.stop();
+            vertx.close();
         }
     }
 
@@ -183,8 +173,7 @@ public final class HelmSliceIT {
      *
      * @since 0.2
      */
-    private static class HelmContainer extends
-        GenericContainer<HelmSliceIT.HelmContainer> {
+    private static class HelmContainer extends GenericContainer<HelmContainer> {
         HelmContainer() {
             super("alpine/helm:2.12.1");
         }
